# Improve to_alternative_strings performance