This commit is the first in a series to improve the performance of
`to_alternative_strings`. I am examining this function as it's a hotspot in file
-> remap -> blackhole performance tests. In this commit I've introduced a
benchmark and have improved performance by reducing vec resize operations
somewhat.

```
path::to_alternative_strings/.foo.(bar | baz)[1].(qux | quux)
                        time:   [681.13 ns 681.86 ns 682.58 ns]
                        thrpt:  [44.709 MiB/s 44.757 MiB/s 44.805 MiB/s]
                 change:
                        time:   [-20.811% -20.567% -20.314%] (p = 0.00 < 0.05)
                        thrpt:  [+25.493% +25.892% +26.280%]
                        Performance has improved.
path::to_alternative_strings/.foo.(bar | baz)[1].(qux | quux).(a|b)
                        time:   [1.3585 us 1.3601 us 1.3618 us]
                        thrpt:  [26.612 MiB/s 26.645 MiB/s 26.676 MiB/s]
                 change:
                        time:   [-12.367% -12.071% -11.613%] (p = 0.00 < 0.05)
                        thrpt:  [+13.139% +13.728% +14.112%]
                        Performance has improved.
path::to_alternative_strings/.foo.(bar | baz)[1].(qux | quux).(a|b).(c|d)
                        time:   [3.4128 us 3.4218 us 3.4312 us]
                        thrpt:  [12.229 MiB/s 12.263 MiB/s 12.295 MiB/s]
                 change:
                        time:   [-3.0548% -2.5211% -2.0710%] (p = 0.00 < 0.05)
                        thrpt:  [+2.1148% +2.5863% +3.1511%]
                        Performance has improved.
```

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

<!--
**Your PR title must conform to the conventional commit spec!**

  <type>!?(<scope>): <description>

  * `type` = chore, enhancement, feat, fix
  * `!` = signals a breaking change
  * `scope` = https://github.com/timberio/vector/blob/master/.github/semantic.yml#L4
  * `description` = short description of the change

Examples:

  * enhancement(file source): Added `sort` option to sort discovered files
  * feat(new source): Initial `statsd` source
  * fix(file source): Fixed a bug discovering new files
  * chore(external docs): Clarified `batch_size` option
-->