@blt 2021-03-22
@jszwedko good question. The initial commit shows that the runtime compared to before the benchmark was introduced, but I didn't make a clean measurement. The initial commit is +20%/+10%/+2% over baseline. Second benchmark is +10%/+20%/+60% over that. Third is +2/+0/+10, fourth is +1/+10/+7. This change ought to be +30/+30/+70 give or take but in the future I'll be careful to get a baseline before making changes. I can't find one in my shell history, but it ought to have been captured in the commit history.